Fold timeout race helpers into timeout files
