+++ Packages/Shared/CmuxAuthRuntime/Sources/CmuxAuthRuntime/Coordinator/AuthPhaseTimeout.swift
@@ -1,5 +1,23 @@
 import Foundation
 
+actor AuthPhaseTimeoutRace {
+    private var hasWinner = false
+
+    func winOperation() -> Bool {
+        win()
+    }
+
+    func winTimeout() -> Bool {
+        win()
+    }
+
+    private func win() -> Bool {
+        guard !hasWinner else { return false }
+        hasWinner = true
+        return true
+    }
+}
+
 /// Race `operation` against a `duration` deadline on `clock`.
 ///
 /// Whichever side finishes first cancels the other and resumes the caller
