--- Packages/iOS/CmuxMobileRPC/Sources/CmuxMobileRPC/RPCTaskTimeoutRace.swift
@@ -1,11 +1,0 @@
-import Foundation
-
-actor RPCTaskTimeoutRace {
-    private var hasWinner = false
-
-    func win() -> Bool {
-        guard !hasWinner else { return false }
-        hasWinner = true
-        return true
-    }
-}
